Agent-Axios/agent-axios-backend/requirements.txt
flask==3.0.0
flask-socketio==5.3.5
eventlet==0.35.2
flask-cors==4.0.0
sqlalchemy==2.0.23
alembic==1.13.1
faiss-cpu==1.9.0
python-dotenv==1.0.0
pydantic==2.5.3
pandas==2.1.4
# Python 3.12 needs the last NumPy release that still ships numpy.distutils for faiss
numpy==1.26.4
requests==2.31.0
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.1
flake8==6.1.0
gitpython==3.1.40
openai>=2.0.0
langsmith
langchain
langchain-core
langchain-openai
langchain-community
langgraph
reportlab==4.0.7
python-socketio[client]==5.10.0
pymilvus>=2.3.0
google-generativeai>=0.3.0
PyJWT==2.8.0
werkzeug>=3.0.0
rich>=13.0.0
